refactor(evals): prepare backfills before batch creation
